arbu/requirements.txt
# Django==5.1
django-bootstrap==0.2.4
# django-bootstrap==24.3
pillow==10.4.0
crispy-bootstrap5==2024.2
django-crispy-forms==2.3
pillow==10.4.0
openpyxl==3.1.5
PyMySQL==1.1.1
mysqlclient==2.2.4
django-environ==0.11.2